REFACTOR: Move test for FullTextSearch to own class
because I plan to add more tests with only subtle differences in the query to test the performance of the method.
Bonus: The other tests are not database-specific and can be run with the "dev" profile once again
