+++ src/test/java/mops/module/services/SuchServiceTest.java
@@ -1,6 +1,5 @@
 package mops.module.services;
 
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.util.List;
@@ -11,9 +10,10 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
 
 @SpringBootTest
-//@ActiveProfiles("dev")
+@ActiveProfiles("dev")
 public class SuchServiceTest {
 
     @Autowired
@@ -74,33 +74,5 @@ public class SuchServiceTest {
     void unsuccessfulSearchReturnsEmptyList() {
         List<Modul> results = suchService.searchForModuleByTitle("katze");
         assertTrue(results.isEmpty());
-    }
-
-    @Test
-    void fullTextSearchTest() {
-        String completeModul = "{'titelDeutsch':'Betriebssysteme','titelEnglisch':'Operating systems',"
-                + "'veranstaltungen':[{'titel':'Vorlesung Betriebssysteme','leistungspunkte':'10CP'"
-                + ",'veranstaltungsformen':[{'form':'Vorlesung','semesterWochenStunden':4},"
-                + "{'form':'Übung','semesterWochenStunden':2}],"
-                + "'beschreibung':{'inhalte':' monolitisch, geschichtet, Mikrokern, Client Server Semaphore, klassische Problemstellungen, Verklemmungen',"
-                + "'lernergebnisse':'Synchronisierung',"
-                + "'literatur':['Alter Schinken'],'verwendbarkeit':['Überall verwendbar'],"
-                + "'voraussetzungenBestehen':['50% der Punkte in der Klausur'],"
-                + "'haeufigkeit':'Alle 2 Semester','sprache':'Deutsch'},"
-                + "'voraussetzungenTeilnahme':['Informatik I'],"
-                + "'zusatzfelder':[{'titel':'Zusatzfeld2',"
-                + "'inhalt':'Dies hier ist das zweite Zusatzfeld!'},"
-                + "{'titel':'Zusatzfeld1','inhalt':'Dies hier ist das erste Zusatzfeld!'}]}],"
-                + "'modulbeauftragte':['Michael Schöttner'],'gesamtLeistungspunkte':'10CP',"
-                + "'studiengang':'Informatik','modulkategorie':'WAHLPFLICHT_BA'}";
-
-        Modul complete = JsonService.jsonObjectToModul(completeModul);
-        complete.refreshMapping();
-        modulRepo.save(complete);
-        List<Modul> results = suchService.searchInVeranstaltungsbeschreibung("problemstellung");
-
-        assertTrue(!results.isEmpty());
-        assertThat(results.get(0).getTitelDeutsch().equals(complete.getTitelDeutsch()));
-
     }
 }
